refactor(ledger): more logical naming/grouping, test organisation
