+++ package.json
@@ -196,6 +196,7 @@
     "download": "8.0.0",
     "esbuild": "0.13.9",
     "esbuild-loader": "2.16.0",
+    "eslint-plugin-deprecation": "1.2.1",
     "eslint-plugin-react-hooks": "4.2.0",
     "generate-json-webpack-plugin": "2.0.0",
     "glob-parent": "6.0.2",
