+++ src/common/stacks-utils.spec.ts
