+++ src/pages/sign-transaction/hooks/use-signed-transaction-fee.ts
@@ -1,5 +1,6 @@
 import { useSignedTransaction } from '@store/transactions/transaction.hooks';
 
+/** @deprecated */
 export function useSignedTransactionFee() {
   const signedTx = useSignedTransaction();
